Update .swiftlint.yml rules
