+++ .swiftlint.yml
@@ -16,6 +16,7 @@ opt_in_rules:
   - closure_spacing
   - collection_alignment
   - comma_inheritance
+  - conditional_returns_on_newline
   - contains_over_filter_count
   - contains_over_filter_is_empty
   - contains_over_first_not_nil
@@ -30,6 +31,8 @@ opt_in_rules:
   - enum_case_associated_values_count
   - expiring_todo
   - explicit_init
+  - fallthrough
+  - fatal_error_message
   - file_name_no_space
   - first_where
   - flatmap_over_map_reduce
@@ -58,6 +61,7 @@ opt_in_rules:
   - override_in_extension
   - pattern_matching_keywords
   - period_spacing
+  - prefer_self_in_static_references
   - prefer_self_type_over_type_of_self
   - prefer_zero_over_explicit_init
   - prefixed_toplevel_constant
@@ -69,6 +73,7 @@ opt_in_rules:
   - required_enum_case
   - return_value_from_void_function
   - self_binding
+  - shorthand_optional_binding
   - sorted_first_last
   - static_operator
   - toggle_bool
@@ -137,7 +142,6 @@ opt_in_rules:
   # - no_extension_access_modifier # (extension_access_modifier)
 
   # Нужен рефакторинг
-  - fatal_error_message
   - file_header
   - file_name
   # - file_types_order
@@ -147,7 +151,6 @@ opt_in_rules:
   - private_action
   - private_outlet
   - redundant_self_in_closure
-  - shorthand_optional_binding
   - sorted_imports
   - strict_fileprivate
   - strong_iboutlet
@@ -155,8 +158,6 @@ opt_in_rules:
 
   # Можно использовать в фреймворках
   # - missing_docs
-
-
 
 type_contents_order:
   order:
@@ -168,49 +169,91 @@ type_contents_order:
     - ib_inspectable
     - ib_outlet
     - instance_property
+    - subscript
     - initializer
     - view_life_cycle_method
     - other_method
     - ib_action
-    - subscript
     - subtype
     - deinitializer
 
 trailing_whitespace:
   ignores_empty_lines: true
 
-implicit_return:
-  included:
-    - closure
-    - getter
-
 indentation_width:
   include_comments: false
   include_multiline_strings: false
+  severity: error
 
 statement_position:
   statement_mode: uncuddled_else
+  severity: error
 
 line_length:
   warning: 180
   ignores_urls: true
+  ignores_comments: true
+
+redundant_void_return:
+  include_closures: false
 
 type_body_length: 
-  warning: 225
+  warning: 350
+  error: 350
 
 nesting:
-  type_level:
-    warning: 5
+  function_level: 1
+  type_level: 5 # Можно поднять, если есть проблема с енамами неймспейсами
 
 identifier_name:
-  min_length: 1
-  max_length: 65
+  min_length:
+    warning: 1
+    error: 1
+  max_length:
+    warning: 50
+    error: 60
 
 type_name:
-  min_length: 3
-  max_length: 65
-  excluded: 
-    - ID
-    - Ad
+  min_length:
+    warning: 2
+    error: 2
+  max_length:
+    warning: 65
+    error: 65
 
-large_tuple: 4
+generic_type_name:
+  max_length:
+    warning: 25
+    error: 30
+
+large_tuple:
+  warning: 4
+  error: 4
+
+cyclomatic_complexity:
+  ignores_case_statements: true
+
+function_body_length:
+  warning: 120
+  error: 200
+
+closure_body_length:
+  warning: 50
+  error: 60
+
+enum_case_associated_values_count:
+  warning: 6
+  error: 6
+
+file_length:
+  warning: 500
+  error: 600
+
+function_parameter_count:
+  warning: 8
+  error: 10
+
+opening_brace:
+  ignore_multiline_type_headers: true
+  ignore_multiline_statement_conditions: true
+  ignore_multiline_function_signatures: true
