+++ Sources/SURCore/Explorer.swift
@@ -55,7 +55,7 @@ public final class Explorer {
         print("🦒 Complete".bold)
     }
     
-    // swiftlint:disable:next cyclomatic_complexity function_body_length
+    // swiftlint:disable:next cyclomatic_complexity
     private func analyze() async throws {
         let exploredResources = await storage.exploredResources
         let exploredUsages = await storage.exploredUsages
