+++ Sources/SURCore/Utils/Path+Utils.swift
@@ -16,7 +16,10 @@ extension Path {
         }
         else {
             // Skip hidden files
-            if lastComponent.hasPrefix(".") { return 0 }
+            if lastComponent.hasPrefix(".") {
+                return 0
+            }
+            
             let attr = try? FileManager.default.attributesOfItem(atPath: absolute().string)
             if let num = attr?[.size] as? NSNumber {
                 return num.intValue
