+++ Sources/SURCore/Utils/String+Utils.swift
@@ -9,8 +9,6 @@ import Foundation
 
 extension String {
     func count(of needle: Character) -> Int {
-        return reduce(0) {
-            $1 == needle ? $0 + 1 : $0
-        }
+        reduce(0) { $1 == needle ? $0 + 1 : $0 }
     }
 }
